requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
alembic==1.14.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
email-validator==2.2.0
jwcrypto==1.5.6
bcrypt==4.2.1
requests==2.32.3
pydantic==2.10.3
python-dotenv==1.0.1
cryptography==44.0.0